renamed auth midlleware, stripped old parts from it
